Extract squashfs utilities to separate module
Refactored squashfs-related code into a dedicated lib/squashfs.py module:
- Moved SquashfsEntry dataclass and parse_unsquashfs_line function
- Moved verify_squashfs_contents function
- Updated imports in ce_install.py
- Renamed test file to test_squashfs.py with updated imports

This improves code organization and makes the squashfs utilities
reusable for future CEFS work.
